// tests/pic/declarations/nnsvector/nnsvector.did.d.ts
export interface Account {
  'owner' : Principal,
  'subaccount' : [] | [Uint8Array | number[]],
}
// ...
export type BillingTransactionFee = { 'none' : null } |
  { 'transaction_percentage_fee_e8s' : bigint } |
  { 'flat_fee_multiplier' : bigint };
// ...
export interface Controller {
  'owner' : Principal,
  'subaccount' : [] | [Uint8Array | number[]],
}
// ...
export interface DestinationEndpointResp {
  'endpoint' : EndpointOpt,
  'name' : string,
}
export type DissolveDelay = { 'Default' : null } |
  { 'DelayDays' : bigint };
export type DissolveStatus = { 'Locked' : null } |
  { 'Dissolving' : null };
export type Endpoint = { 'ic' : EndpointIC } |
  { 'other' : EndpointOther };
export interface EndpointIC { 'ledger' : Principal, 'account' : Account }
export type EndpointIdx = number;
export type EndpointOpt = { 'ic' : EndpointOptIC } |
  { 'other' : EndpointOptOther };
// ...
export type LocalNodeId = number;
// ...
export interface NodeShared {
  'id' : LocalNodeId,
  'created' : bigint,
  'active' : boolean,
  'modified' : bigint,
  'controllers' : Array<Controller>,
  'custom' : [] | [Shared],
  'extractors' : Uint32Array | number[],
  'billing' : {
    'transaction_fee' : BillingTransactionFee,
    'expires' : [] | [bigint],
    'current_balance' : bigint,
    'billing_option' : bigint,
    'account' : Account,
    'frozen' : boolean,
    'cost_per_day' : bigint,
  },
  'destinations' : Array<DestinationEndpointResp>,
  'sources' : Array<SourceEndpointResp>,
  'refund' : Account,
}
export interface PylonMetaResp {
  'name' : string,
  'billing' : BillingPylon,
  'supported_ledgers' : Array<SupportedLedger>,
  'request_max_expire_sec' : bigint,
  'governed_by' : string,
  'temporary_nodes' : { 'allowed' : boolean, 'expire_sec' : bigint },
  'modules' : Array<ModuleMeta>,
}
export type Shared = { 'devefi_jes1_icpneuron' : Shared__1 };
// ...
export interface Shared__1 {
  'log' : Array<Activity>,
  'internals' : {
    'local_idx' : number,
    'refresh_idx' : [] | [bigint],
    'updating' : UpdatingStatus,
    'spawning_neurons' : Array<SharedNeuronCache>,
  },
  'cache' : SharedNeuronCache,
  'variables' : {
    'dissolve_delay' : DissolveDelay,
    'dissolve_status' : DissolveStatus,
    'followee' : Followee,
  },
}
export interface SourceEndpointResp {
  'balance' : bigint,
  'endpoint' : Endpoint,
  'name' : string,
}